Added Tombstone and Fixed Some Bugs with Attack
Bug 1:  Attack's trigger weren't being detected by tombstone.
2 reasons:
disabling and reenabling collider counts as an on enter event, but disabling + reenabling entire gameObject does not.
There is a bug with Unity where if you want the above behaviour to happen IN CODE, the collider has to be moving, or trigger events will not occur.  Turning collider on and off in the inspector works without this additional step.

Bug 2:  Player attack animation and attack collider spawn are not insync
Fix: Attack gets called in playercontroller, instead of being called separately using button.